+++ lib/graphql/get_app_by_api_key.graphql
