+++ lib/project_types/extension/forms/create.rb
@@ -3,9 +3,12 @@
 module Extension
   module Forms
     class Create < ShopifyCli::Form
-      flag_arguments :name, :type
+      flag_arguments :name, :type, :api_key
+
+      attr_reader :app
 
       def ask
+        self.app = ask_app
         self.type = ask_type
         self.name = ask_name
       end
@@ -15,6 +18,36 @@ module Extension
       end
 
       private
+
+      attr_writer :app
+
+      def ask_app
+        if !api_key.nil?
+          found_app = Tasks::GetApp.call(context: ctx, api_key: api_key)
+          ctx.abort(ctx.message("register.invalid_api_key", api_key)) if found_app.nil?
+          found_app
+        else
+          apps = load_apps
+          CLI::UI::Prompt.ask(ctx.message("register.ask_app")) do |handler|
+            apps.each do |app|
+              handler.option("#{app.title} by #{app.business_name}") { app }
+            end
+          end
+        end
+      end
+
+      def load_apps
+        ctx.puts(@ctx.message("register.loading_apps"))
+        apps = Tasks::GetApps.call(context: ctx)
+
+        apps.empty? ? abort_no_apps : apps
+      end
+
+      def abort_no_apps
+        ctx.puts(@ctx.message("register.no_apps"))
+        ctx.puts(@ctx.message("register.learn_about_apps"))
+        raise ShopifyCli::AbortSilent
+      end
 
       def ask_name
         ask_with_reprompt(
